requirements.txt
Kivy==2.2.1
Kivy-garden-draggable==0.2.0